backend/requirements.txt
# Django>=5.0.0
Django
djangorestframework==3.15.1
django-cors-headers==4.3.1
# psycopg2==2.9.9